+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Fixed
 - Fixed the agent `--annotations` and `--labels` flags.
 
+### Added
+- Added the message bus to Tessend in order to track Tessen configuration changes from the API.
+
 ## [5.5.1] - 2019-04-15
 
 ### Changed